requirements.txt
aiofiles==23.2.1
aioredis==2.0.1
aiosqlite==0.17.0
amqp==5.1.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
Babel==2.13.1
bcrypt==4.0.1
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.4.2
e==1.4.5
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.2
fastapi-admin==1.0.4
flake8==6.1.0
flower==2.0.1
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
humanize==4.8.0
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.6.1
mypy-extensions==1.0.0
orjson==3.9.8
pendulum==2.1.2
prometheus-client==0.18.0
prompt-toolkit==3.0.39
pyasn1==0.5.1
pycodestyle==2.11.1
pydantic==2.4.2
pydantic-core==2.10.1
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pyflakes==3.1.0
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
tornado==6.3.3
tortoise-orm==0.20.0
typing-extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.6
uvicorn==0.23.2
vine==5.0.0
watchfiles==0.20.0
wcwidth==0.2.9
websockets==11.0.3
git+https://github.com/sepehrsafa/owjcommon#egg=owjcommon
